AgriVisionAI-main/tts_service/requirements.txt
Flask==2.3.3
python-dotenv
google-cloud-texttospeech
